Drastically improved ROM access times
This uses a custom filesystem format called MMFS (memory-mapped filesystem),
which maps the entire partition into memory and uses direct memory addresses
to read files. This is a lot more performant than littlefs, as it's optimized for
read-only filesystems and doesn't have to go through the flash peripheral;
but it's read-only (which is fine for the ROM).